OSX cross building (experiment) (#7)
* handle osx cross building (wip)

* refine cross-builds

* enable artifacts & rerender

MNT: Re-rendered with conda-build 3.22.0, conda-smithy 3.22.1, and conda-forge-pinning 2022.12.29.19.05.49

* ensure dir exists

* only test on native and emulated

MNT: Re-rendered with conda-build 3.22.0, conda-smithy 3.22.1, and conda-forge-pinning 2023.01.02.15.58.50

* less intrusive

+++ recipe/build.sh
@@ -10,13 +10,16 @@ done
 # make sure pyinstaller finds Apple's codesign first in PATH
 # some base installations have 'sigtool', which ships a
 # 'codesign' binary that might shadow Apple's codesign
-if [[ $target_platform == osx-* ]]; then
+if [[ $target_platform == osx-* && ! -f "$BUILD_PREFIX/bin/codesign" ]]; then
+  mkdir -p "$BUILD_PREFIX/bin"
   ln -s /usr/bin/codesign "$BUILD_PREFIX/bin/codesign"
 fi
 
-# -F is to create a single file
-# -s strips executables and libraries
-pyinstaller conda.exe.spec
+# when cross-building, we need to run build's python (but without the cross-python magic)
+if [[ $build_platform != $target_platform ]]; then  
+  export PYTHON="$BUILD_PREFIX/bin/python"
+fi
+"${PYTHON}" -m PyInstaller conda.exe.spec
 mkdir -p "$PREFIX/standalone_conda"
 mv dist/conda.exe "$PREFIX/standalone_conda"
 # clean up .pyc files that pyinstaller creates
